Added link to issue to test, removed unused import
