+++ prusti-common/src/vir/ast/expr.rs
@@ -4,8 +4,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use viper::AstFactory;
-
 use super::super::borrows::Borrow;
 use crate::vir::{ast::*, FloatSize::*};
 use std::collections::{HashMap, HashSet};
