+++ prusti-tests/tests/verify/fail/floats/nan-cmp-issues/nan_eq_nan_f64.rs
@@ -1,5 +1,4 @@
 // ignore-test: https://github.com/viperproject/prusti-dev/issues/575
 fn main() {
     assert!(f64::NAN == f64::NAN); //~ ERROR: the asserted expression might not hold
-    // assert!(f64::NAN != f64::NAN);
 }
