+++ prusti-tests/tests/verify/fail/floats/nan_cmp_nan_f32.rs
@@ -1,0 +1,20 @@
+fn eq32() { 
+    assert!(f32::NAN == f32::NAN); //~ ERROR: the asserted expression might not hold
+} 
+fn ne32() { 
+    assert!(f32::NAN != f32::NAN);
+}
+fn leq32() { 
+    assert!(f32::NAN <= f32::NAN); //~ ERROR: the asserted expression might not hold
+} 
+fn lt32() { 
+    assert!(f32::NAN < f32::NAN); //~ ERROR: the asserted expression might not hold
+}
+fn geq32() { 
+    assert!(f32::NAN >= f32::NAN); //~ ERROR: the asserted expression might not hold
+} 
+fn gt32() { 
+    assert!(f32::NAN > f32::NAN); //~ ERROR: the asserted expression might not hold
+}
+
+fn main() {}
