+++ prusti-tests/tests/verify/fail/floats/nan_cmp_nan_f64.rs
@@ -1,23 +1,4 @@
 // ignore-test: https://github.com/viperproject/prusti-dev/issues/575
-fn eq32() { 
-    assert!(f32::NAN == f32::NAN); //~ ERROR: the asserted expression might not hold
-} 
-fn ne32() { 
-    assert!(f32::NAN != f32::NAN);
-}
-fn leq32() { 
-    assert!(f32::NAN <= f32::NAN); //~ ERROR: the asserted expression might not hold
-} 
-fn lt32() { 
-    assert!(f32::NAN < f32::NAN); //~ ERROR: the asserted expression might not hold
-}
-fn geq32() { 
-    assert!(f32::NAN >= f32::NAN); //~ ERROR: the asserted expression might not hold
-} 
-fn gt32() { 
-    assert!(f32::NAN > f32::NAN); //~ ERROR: the asserted expression might not hold
-}
-
 fn eq64() { 
     assert!(f64::NAN == f64::NAN); //~ ERROR: the asserted expression might not hold
 }
